Fix test to use appropriate composition
